Trying some fixes.
Fixes for build, not for test.

Undoing some changes but keeping the optimized parts.

Trying to get test failure for namespace output in actions.
