+++ src/message_buffer.cpp
@@ -1,4 +1,5 @@
 #include <quicr/message_buffer.h>
+#include <quicr/uvarint.h>
 
 #include <algorithm>
 #include <array>
@@ -139,92 +140,6 @@ operator>>(MessageBuffer& msg, std::vector<uint8_t>& val)
   msg >> vec_size;
 
   val = msg.pop_front(vec_size);
-  return msg;
-}
-
-MessageBuffer&
-operator<<(MessageBuffer& msg, const uintVar_t& v)
-{
-  uint64_t val = v;
-
-  if (val < ((uint64_t)1 << 7)) {
-    msg.push(uint8_t(((val >> 0) & 0x7F)) | 0x00);
-    return msg;
-  }
-
-  if (val < ((uint64_t)1 << 14)) {
-    msg.push(uint8_t(((val >> 8) & 0x3F) | 0x80));
-    msg.push(uint8_t((val >> 0) & 0xFF));
-    return msg;
-  }
-
-  if (val < ((uint64_t)1 << 29)) {
-    msg.push(uint8_t(((val >> 24) & 0x1F) | 0x80 | 0x40));
-    msg.push(uint8_t((val >> 16) & 0xFF));
-    msg.push(uint8_t((val >> 8) & 0xFF));
-    msg.push(uint8_t((val >> 0) & 0xFF));
-    return msg;
-  }
-
-  msg.push(uint8_t(((val >> 56) & 0x0F) | 0x80 | 0x40 | 0x20));
-  msg.push(uint8_t((val >> 48) & 0xFF));
-  msg.push(uint8_t((val >> 40) & 0xFF));
-  msg.push(uint8_t((val >> 32) & 0xFF));
-  msg.push(uint8_t((val >> 24) & 0xFF));
-  msg.push(uint8_t((val >> 16) & 0xFF));
-  msg.push(uint8_t((val >> 8) & 0xFF));
-  msg.push(uint8_t((val >> 0) & 0xFF));
-
-  return msg;
-}
-
-MessageBuffer&
-operator>>(MessageBuffer& msg, uintVar_t& v)
-{
-  uint8_t byte[8] = { 0, 0, 0, 0, 0, 0, 0, 0 };
-  uint8_t first = msg.front();
-
-  if ((first & (0x80)) == 0) {
-    msg >> byte[0];
-    uint8_t val = ((byte[0] & 0x7F) << 0);
-    v = val;
-    return msg;
-  }
-
-  if ((first & (0x80 | 0x40)) == 0x80) {
-    msg >> byte[1];
-    msg >> byte[0];
-    uint16_t val = (((uint16_t)byte[1] & 0x3F) << 8) + ((uint16_t)byte[0] << 0);
-    v = val;
-    return msg;
-  }
-
-  if ((first & (0x80 | 0x40 | 0x20)) == (0x80 | 0x40)) {
-    msg >> byte[3];
-    msg >> byte[2];
-    msg >> byte[1];
-    msg >> byte[0];
-    uint32_t val = ((uint32_t)(byte[3] & 0x1F) << 24) +
-                   ((uint32_t)byte[2] << 16) + ((uint32_t)byte[1] << 8) +
-                   ((uint32_t)byte[0] << 0);
-    v = val;
-    return msg;
-  }
-
-  msg >> byte[7];
-  msg >> byte[6];
-  msg >> byte[5];
-  msg >> byte[4];
-  msg >> byte[3];
-  msg >> byte[2];
-  msg >> byte[1];
-  msg >> byte[0];
-  uint64_t val = ((uint64_t)(byte[7] & 0x0F) << 56) +
-                 ((uint64_t)(byte[6]) << 48) + ((uint64_t)(byte[5]) << 40) +
-                 ((uint64_t)(byte[4]) << 32) + ((uint64_t)(byte[3]) << 24) +
-                 ((uint64_t)(byte[2]) << 16) + ((uint64_t)(byte[1]) << 8) +
-                 ((uint64_t)(byte[0]) << 0);
-  v = val;
   return msg;
 }
